src/ContosoSuitesDashboard/requirements.txt
pandas
numpy
scikit-learn
streamlit
seaborn
plotly.express
openai